@@ -1,6 +1,7 @@
 {
   "author": "Jake Luer <jake@alogicalparadox.com>",
   "name": "chai",
+  "type": "module",
   "description": "BDD/TDD assertion library for node.js and the browser. Test framework agnostic.",
   "keywords": [
     "test",
@@ -17,7 +18,7 @@
     "Veselin Todorov <hi@vesln.com>",
     "John Firebaugh <john.firebaugh@gmail.com>"
   ],
-  "version": "4.4.1",
+  "version": "5.1.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/chaijs/chai"
@@ -25,39 +26,33 @@
   "bugs": {
     "url": "https://github.com/chaijs/chai/issues"
   },
-  "main": "./index",
-  "exports": {
-    ".": {
-      "require": "./index.js",
-      "import": "./index.mjs"
-    },
-    "./*": "./*"
-  },
+  "main": "./chai.js",
   "scripts": {
-    "test": "make test"
+    "prebuild": "npm run clean",
+    "build": "npm run build:esm",
+    "build:esm": "esbuild --bundle --format=esm --keep-names --outfile=chai.js index.js",
+    "pretest": "npm run build",
+    "test": "npm run test-node && npm run test-chrome",
+    "test-node": "mocha --require ./test/bootstrap/index.js --reporter dot test/*.js",
+    "test-chrome": "web-test-runner --playwright",
+    "clean": "rm -f chai.js coverage"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=12"
   },
   "dependencies": {
-    "assertion-error": "^1.1.0",
-    "check-error": "^1.0.3",
-    "deep-eql": "^4.1.3",
-    "get-func-name": "^2.0.2",
-    "loupe": "^2.3.6",
-    "pathval": "^1.1.1",
-    "type-detect": "^4.0.8"
+    "assertion-error": "^2.0.1",
+    "check-error": "^2.0.0",
+    "deep-eql": "^5.0.1",
+    "loupe": "^3.1.0",
+    "pathval": "^2.0.0"
   },
   "devDependencies": {
-    "browserify": "^16.5.2",
-    "bump-cli": "^2.7.1",
-    "codecov": "^3.8.3",
-    "istanbul": "^0.4.5",
-    "karma": "^6.4.2",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-firefox-launcher": "^1.3.0",
-    "karma-mocha": "^2.0.1",
-    "karma-sauce-launcher": "^4.1.4",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@web/dev-server-rollup": "^0.6.1",
+    "@web/test-runner": "^0.18.0",
+    "@web/test-runner-playwright": "^0.11.0",
+    "esbuild": "^0.19.10",
     "mocha": "^10.2.0"
   }
 }
