@@ -1,45 +1,42 @@
 {
-  "name": "is-fullwidth-code-point",
-  "version": "2.0.0",
-  "description": "Check if the character represented by a given Unicode code point is fullwidth",
-  "license": "MIT",
-  "repository": "sindresorhus/is-fullwidth-code-point",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "fullwidth",
-    "full-width",
-    "full",
-    "width",
-    "unicode",
-    "character",
-    "char",
-    "string",
-    "str",
-    "codepoint",
-    "code",
-    "point",
-    "is",
-    "detect",
-    "check"
-  ],
-  "devDependencies": {
-    "ava": "*",
-    "xo": "*"
-  },
-  "xo": {
-    "esnext": true
-  }
+	"name": "is-fullwidth-code-point",
+	"version": "3.0.0",
+	"description": "Check if the character represented by a given Unicode code point is fullwidth",
+	"license": "MIT",
+	"repository": "sindresorhus/is-fullwidth-code-point",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "sindresorhus.com"
+	},
+	"engines": {
+		"node": ">=8"
+	},
+	"scripts": {
+		"test": "xo && ava && tsd-check"
+	},
+	"files": [
+		"index.js",
+		"index.d.ts"
+	],
+	"keywords": [
+		"fullwidth",
+		"full-width",
+		"full",
+		"width",
+		"unicode",
+		"character",
+		"string",
+		"codepoint",
+		"code",
+		"point",
+		"is",
+		"detect",
+		"check"
+	],
+	"devDependencies": {
+		"ava": "^1.3.1",
+		"tsd-check": "^0.5.0",
+		"xo": "^0.24.0"
+	}
 }
