@@ -1,7 +1,7 @@
 {
   "name": "workerpool",
   "license": "Apache-2.0",
-  "version": "6.1.0",
+  "version": "6.2.1",
   "description": "Offload tasks to a pool of workers on node.js and in the browser",
   "homepage": "https://github.com/josdejong/workerpool",
   "author": "Jos de Jong <wjosdejong@gmail.com> (https://github.com/josdejong)",
@@ -34,19 +34,19 @@
     "prepublishOnly": "npm run test"
   },
   "devDependencies": {
-    "@babel/core": "7.12.10",
-    "@babel/preset-env": "7.12.11",
-    "babel-loader": "8.2.2",
-    "date-format": "3.0.0",
+    "@babel/core": "7.17.9",
+    "@babel/preset-env": "7.16.11",
+    "babel-loader": "8.2.4",
+    "date-format": "4.0.6",
     "del": "6.0.0",
-    "fancy-log": "1.3.3",
-    "find-process": "1.4.4",
+    "fancy-log": "2.0.0",
+    "find-process": "1.4.7",
     "gulp": "4.0.2",
-    "handlebars": "4.7.6",
+    "handlebars": "4.7.7",
     "istanbul": "0.4.5",
-    "mocha": "8.2.1",
-    "uglify-js": "3.12.5",
-    "webpack": "5.19.0"
+    "mocha": "9.2.2",
+    "uglify-js": "3.15.4",
+    "webpack": "5.72.0"
   },
   "dependencies": {}
 }
